Introducing IDisplayManager<T> and IDisplayDriver<T> for reusability
