+++ src/OrchardCore.Modules/Orchard.Settings/Drivers/DefaultSiteSettingsDisplayDriver.cs
@@ -1,38 +1,39 @@
 ﻿using System;
 using System.Threading.Tasks;
 using Orchard.DisplayManagement.Handlers;
-using Orchard.DisplayManagement.ModelBinding;
 using Orchard.DisplayManagement.Views;
-using Orchard.Settings.Services;
 using Orchard.Settings.ViewModels;
 
 namespace Orchard.Settings.Drivers
 {
-    public class DefaultSiteSettingsDisplayDriver : SiteSettingsDisplayDriver
+    public class DefaultSiteSettingsDisplayDriver : DisplayDriver<ISite>
     {
-        public override IDisplayResult Edit(ISite site, BuildEditorContext context)
-        {
-            return Combine(
-                Shape<SiteSettingsViewModel>("SiteSettings_Edit", model =>
-                {
-                    model.SiteName = site.SiteName;
-                    model.TimeZone = site.TimeZone;
-                    model.TimeZones = TimeZoneInfo.GetSystemTimeZones();
-                }).Location("Content:1").OnGroup("general"),
+        public const string GroupId = "general";
 
-                Shape("SiteSettings_SaveButton")
-                    .Location("Actions")
-                    .OnGroup(context.GroupId) // Trick to render the shape for all groups
+        public override Task<IDisplayResult> EditAsync(ISite site, BuildEditorContext context)
+        {
+            return Task.FromResult<IDisplayResult>(
+                Combine(
+                    Shape<SiteSettingsViewModel>("SiteSettings_Edit", model =>
+                    {
+                        model.SiteName = site.SiteName;
+                        model.TimeZone = site.TimeZone;
+                        model.TimeZones = TimeZoneInfo.GetSystemTimeZones();
+                    }).Location("Content:1").OnGroup(GroupId),
+
+                    Shape("SiteSettings_SaveButton")
+                        .Location("Actions")
+                        .OnGroup(context.GroupId) // Trick to render the shape for all groups
+                )
             );
         }
-
-        public override async Task<IDisplayResult> UpdateAsync(ISite site, IUpdateModel updater, string groupId)
+        public override async Task<IDisplayResult> UpdateAsync(ISite site, UpdateEditorContext context)
         {
-            if (groupId == "general")
+            if (context.GroupId == GroupId)
             {
                 var model = new SiteSettingsViewModel();
 
-                if (await updater.TryUpdateModelAsync(model, Prefix, t => t.SiteName, t => t.TimeZone))
+                if (await context.Updater.TryUpdateModelAsync(model, Prefix, t => t.SiteName, t => t.TimeZone))
                 {
                     site.SiteName = model.SiteName;
                     site.TimeZone = model.TimeZone;
