--- src/OrchardCore.Modules/Orchard.Settings/Services/ISiteSettingsDisplayHandler.cs
@@ -1,11 +1,0 @@
-﻿using System.Threading.Tasks;
-using Orchard.DisplayManagement.Handlers;
-
-namespace Orchard.Settings.Services
-{
-    public interface ISiteSettingsDisplayHandler
-    {
-        Task BuildEditorAsync(ISite site, BuildEditorContext context);
-        Task UpdateEditorAsync(ISite site, UpdateEditorContext context);
-    }
-}
