+++ src/OrchardCore/Orchard.DisplayManagement/DisplayManager.cs
@@ -1,63 +1,132 @@
-﻿using System.Threading.Tasks;
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Modules;
+using Microsoft.Extensions.Logging;
 using Orchard.DisplayManagement.Descriptors;
 using Orchard.DisplayManagement.Handlers;
+using Orchard.DisplayManagement.Layout;
+using Orchard.DisplayManagement.ModelBinding;
 using Orchard.DisplayManagement.Theming;
-using Orchard.DisplayManagement.Zones;
 
 namespace Orchard.DisplayManagement
 {
-    public abstract class BaseDisplayManager
+    public class DisplayManager<TModel> : BaseDisplayManager, IDisplayManager<TModel>
     {
+        private static readonly string TypeName = typeof(TModel).Name;
+        private static readonly string EditShapeType = typeof(TModel).Name + "_Edit";
+
+        private readonly IEnumerable<IDisplayDriver<TModel>> _drivers;
         private readonly IShapeTableManager _shapeTableManager;
         private readonly IShapeFactory _shapeFactory;
         private readonly IThemeManager _themeManager;
+        private readonly ILayoutAccessor _layoutAccessor;
 
-        public BaseDisplayManager(
+        public DisplayManager(
+            IEnumerable<IDisplayDriver<TModel>> drivers,
             IShapeTableManager shapeTableManager,
             IShapeFactory shapeFactory,
-            IThemeManager themeManager
-            )
+            IThemeManager themeManager,
+            ILogger<DisplayManager<TModel>> logger,
+            ILayoutAccessor layoutAccessor
+            ) : base(shapeTableManager, shapeFactory, themeManager)
         {
             _shapeTableManager = shapeTableManager;
             _shapeFactory = shapeFactory;
             _themeManager = themeManager;
+            _layoutAccessor = layoutAccessor;
+            _drivers = drivers;
+
+            Logger = logger;
         }
 
-        protected async Task BindPlacementAsync(IBuildShapeContext context)
+        ILogger Logger { get; set; }
+
+        public async Task<dynamic> BuildDisplayAsync(TModel model, IUpdateModel updater, string displayType = null, string group = null)
         {
-            var theme = await _themeManager.GetThemeAsync();
-            var shapeTable = _shapeTableManager.GetShapeTable(theme.Id);
+            var actualShapeType = TypeName;
+            var actualDisplayType = string.IsNullOrEmpty(displayType) ? "Detail" : displayType;
 
-            context.FindPlacement = (shapeType, differentiator, displayType, displayContext) => FindPlacementImpl(shapeTable, shapeType, differentiator, displayType, context);
-        }
-
-        private static PlacementInfo FindPlacementImpl(ShapeTable shapeTable, string shapeType, string differentiator, string displayType, IBuildShapeContext context)
-        {
-            ShapeDescriptor descriptor;
-
-            if (shapeTable.Descriptors.TryGetValue(shapeType, out descriptor))
+            // _[DisplayType] is only added for the ones different than Detail
+            if (actualDisplayType != "Detail")
             {
-                var placementContext = new ShapePlacementContext(
-                    shapeType,
-                    displayType,
-                    differentiator,
-                    context.Shape
-                );
-
-                var placement = descriptor.Placement(placementContext);
-                if (placement != null)
-                {
-                    placement.Source = placementContext.Source;
-                    return placement;
-                }
+                actualShapeType = actualShapeType + "_" + actualDisplayType;
             }
 
-            return null;
+            dynamic shape = CreateContentShape(actualShapeType);
+
+            var context = new BuildDisplayContext(
+                shape,
+                actualDisplayType,
+                group ?? "",
+                _shapeFactory,
+                _layoutAccessor.GetLayout(),
+                updater
+            );
+
+            await BindPlacementAsync(context);
+
+            await _drivers.InvokeAsync(async driver =>
+            {
+                var result = await driver.BuildDisplayAsync(model, context);
+                if (result != null)
+                    result.Apply(context);
+            }, Logger);
+
+            return shape;
         }
 
-        protected dynamic CreateContentShape(string actualShapeType)
+        public async Task<dynamic> BuildEditorAsync(TModel model, IUpdateModel updater, string group = null)
         {
-            return _shapeFactory.Create(actualShapeType, () => new ZoneHolding(() => _shapeFactory.Create("ContentZone", Arguments.Empty)));
+            dynamic shape = CreateContentShape(EditShapeType);
+
+            var context = new BuildEditorContext(
+                shape,
+                group ?? "",
+                "",
+                _shapeFactory,
+                _layoutAccessor.GetLayout(),
+                updater
+            );
+
+            await BindPlacementAsync(context);
+
+            await _drivers.InvokeAsync(async driver =>
+            {
+                var result = await driver.BuildEditorAsync(model, context);
+                if (result != null)
+                {
+                    result.Apply(context);
+                }
+            }, Logger);
+
+            return shape;
+        }
+
+        public async Task<dynamic> UpdateEditorAsync(TModel model, IUpdateModel updater, string group = null)
+        {
+            dynamic shape = CreateContentShape(EditShapeType);
+
+            var context = new UpdateEditorContext(
+                shape,
+                group ?? "",
+                "",
+                _shapeFactory,
+                _layoutAccessor.GetLayout(),
+                updater
+            );
+
+            await BindPlacementAsync(context);
+
+            await _drivers.InvokeAsync(async driver =>
+            {
+                var result = await driver.UpdateEditorAsync(model, context);
+                if (result != null)
+                {
+                    result.Apply(context);
+                }
+            }, Logger);
+
+            return shape;
         }
     }
 }
