+++ src/OrchardCore/Orchard.DisplayManagement/Handlers/IDisplayDriver.cs
@@ -1,4 +1,4 @@
-﻿using Orchard.DisplayManagement.Views;
+using Orchard.DisplayManagement.Views;
 using System.Threading.Tasks;
 
 namespace Orchard.DisplayManagement.Handlers
@@ -11,5 +11,10 @@ namespace Orchard.DisplayManagement.Handlers
         Task<IDisplayResult> BuildDisplayAsync(TModel model, TDisplayContext context);
         Task<IDisplayResult> BuildEditorAsync(TModel model, TEditorContext context);
         Task<IDisplayResult> UpdateEditorAsync(TModel model, TUpdateContext context);
+    }
+
+    public interface IDisplayDriver<TModel> : IDisplayDriver<TModel, BuildDisplayContext, BuildEditorContext, UpdateEditorContext>
+    {
+
     }
 }
