+++ src/OrchardCore/Orchard.DisplayManagement/IDisplayManager.cs
@@ -5,8 +5,8 @@ namespace Orchard.DisplayManagement
 {
     public interface IDisplayManager<TModel>
     {
-        Task<dynamic> BuildDisplayAsync(TModel model, string displayType = "", string groupId = "");
-        Task<dynamic> BuildEditorAsync(TModel model, string groupId = "");
+        Task<dynamic> BuildDisplayAsync(TModel model, IUpdateModel updater, string displayType = "", string groupId = "");
+        Task<dynamic> BuildEditorAsync(TModel model, IUpdateModel updater, string groupId = "");
         Task<dynamic> UpdateEditorAsync(TModel model, IUpdateModel updater, string groupId = "");
     }
 }
